requirements/requirements.txt
pytest
pyusb
numpy
